rename @types to types
