+++ types/StyleDictionaryConfigGenerator.d.ts
